Add files via upload
The is the folder which contains MAC operation for int and cocotb code which checks coverage for int operation of a*b + c. The user can change the range based on the computation resources available.The Virtual Machine resources limit our range of test for coverage. We have covered for a sample range which can be extended!